Test your GPIO pin to maker it's not blown
If you're not getting a Din reading, this is a sanity check that will at least tell you if your pin is noticing more than all zeroes. Maybe it can makes sense of Din but it can still tell you it's alive. This is for GPIO27 but can easily be changed.